[kernel: io] add support for arguments to print_string this commit adds basic support for arguments when printing a string. at the moment you can print strings and unsigned (char, short, int and long)s. a method called toString() has been added as a template. this allows us to do toString(any int type) and it will convert it to a readable string.

+++ src/kernel/interrupt/idt/types/interrupt_frame.h
@@ -1,12 +1,14 @@
 #ifndef KERNEL_INTERRUPT_FRAME_H
 #define KERNEL_INTERRUPT_FRAME_H
 
+#include <stddef.h>
+
 struct interrupt_frame {
-    uint16_t instruction_pointer;
-    uint16_t code_segment;
-    uint16_t flags;
-    uint16_t stack_pointer;
-    uint16_t stack_segment;
+    size_t instruction_pointer;
+    size_t code_segment;
+    size_t flags;
+    size_t stack_pointer;
+    size_t stack_segment;
 };
 
 #endif //KERNEL_INTERRUPT_FRAME_H
